flatimage with wine

+++ src/wine.sh
@@ -11,6 +11,8 @@
 
 set -e
 
+shopt -s nullglob
+
 GIMG_SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 
 source "$GIMG_SCRIPT_DIR/common.sh"
@@ -18,25 +20,40 @@ source "$GIMG_SCRIPT_DIR/common.sh"
 # Get wine appimage and define wine alias
 function _fetch_wine()
 {
-  local url
-
-  url=$("$GIMG_SCRIPT_DIR"/busybox wget --header="Accept: application/vnd.github+json" -O - \
-    https://api.github.com/repos/ruanformigoni/wine/releases 2>&1 |
-    grep -Eo "https://.*continuous-.*/wine-$GIMG_WINE_DIST-[0-9.-]+-continuous-x86_64.AppImage\"" ||
-    die "Error fetching wine url")
+  if [[ "$GIMG_PKG_TYPE" = "flatimage" ]]; then
+    local url="https://gitlab.com/api/v4/projects/45732205/packages"
+    url+="/generic/wine-amd-intel/continuous/wine-amd_intel-$GIMG_WINE_DIST.fim"
+  else
+    url=$("$GIMG_SCRIPT_DIR"/busybox wget --header="Accept: application/vnd.github+json" -O - \
+      https://api.github.com/repos/ruanformigoni/wine/releases 2>&1 |
+      grep -Eo "https://.*continuous-.*/wine-$GIMG_WINE_DIST-[0-9.-]+-continuous-x86_64.AppImage\"" ||
+      die "Error fetching wine url")
+    url="${url%\"}"
+  fi
 
   if [ ! -f "AppDir/usr/bin/wine" ]; then
     if [ -f "./wine" ]; then
       mv wine AppDir/usr/bin/wine
     else
-      _fetch "wine" "${url%\"}"
+      _fetch "wine" "$url"
       mv wine AppDir/usr/bin/wine
     fi
   fi
 
-  if [ ! -f "AppDir/usr/bin/winetricks" ]; then
+  # Create winetricks
+  local path_winetricks="AppDir/usr/bin/winetricks"
+  if [[ "$GIMG_PKG_TYPE" = "flatimage" ]]; then
+    { sed -E 's/^\s+://' | tee "$path_winetricks" &>/dev/null; } <<-"END"
+    :  #!/usr/bin/env bash
+    :
+    :  SCRIPT_DIR="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd)"
+    :
+    :  PATH="$PATH:/opt/wine/bin" "$SCRIPT_DIR"/wine fim-exec winetricks "$@"
+		END
+  else
     ln -sf wine AppDir/usr/bin/winetricks
   fi
+  chmod +x "$path_winetricks"
 }
 
 function arch_select()
@@ -139,12 +156,17 @@ function wine_test()
 
 function wine_executable_select()
 {
-  msg "Install method: $GIMG_PKG_TYPE"
   msg "Select the main executable"
   _eval_select 'find "AppDir/app/wine" -not -path "*drive_c/windows/*.exe" -iname "*.exe"'
-  local executable="$_FN_RET"
-  local dir_build="$1"
-  local name="$2"
+  echo "$_FN_RET"
+}
+
+function wine_package_method()
+{
+  msg "Install method: $GIMG_PKG_METHOD"
+  local executable="$1"
+  local dir_build="$2"
+  local name="$3"
 
   # Get directory to move out from drive c:
   local dir_installation
@@ -152,24 +174,34 @@ function wine_executable_select()
   dir_installation="${dir_installation//*drive_c\/}"
   dir_installation="${dir_installation//\/*}"
 
+  local dir_target="AppDir/app/rom"
+
   # Create directory to store installed files
   # Move to external prefix or keep it inside appimage
-  if [ "${GIMG_PKG_TYPE}" = "overlayfs" ]; then
+  ## Make it writteable with overlayfs
+  if [ "${GIMG_PKG_METHOD}" = "overlayfs" ]; then
     cp "${GIMG_SCRIPT_DIR}/overlayfs" "AppDir/usr/bin"
-  elif [ "${GIMG_PKG_TYPE}" = "unionfs" ]; then
+  ## Make it writteable with unionfs
+  elif [ "${GIMG_PKG_METHOD}" = "unionfs" ]; then
     cp "${GIMG_SCRIPT_DIR}/unionfs" "AppDir/usr/bin"
-  elif [ "${GIMG_PKG_TYPE}" = "readonly" ]; then
-    dir_target="AppDir/app/rom"
+  ## Copy on prefix not found
+  elif [ "${GIMG_PKG_METHOD}" = "copy" ]; then
     mkdir -p "$dir_target"
     # Move installed software to target directory
     msg "Moving '$dir_installation' to '$dir_target'"
     mv "AppDir/app/wine/drive_c/$dir_installation" "$dir_target"
-  else # prefix
-    dir_target="$dir_build/.${name}.AppImage.config"
+  ## Keep prefix as is
+  elif [ "${GIMG_PKG_METHOD}" = "prefix" ]; then
+    dir_target="$dir_build/.${name}.${GIMG_PKG_TYPE}.config"
     mkdir -p "$dir_target"
-    # Move prefix to outside of AppImage
+    # Move prefix
     msg "Moving '$dir_build/AppDir/app/wine' to '$dir_target'"
-    mv "$dir_build/AppDir/app/wine" "$dir_target"
+    mv "$dir_build/AppDir/app/wine" "$dir_target"/wine
+  ## Writteable inside the image (flatimage only)
+  elif [ "${GIMG_PKG_METHOD}" = "dynamic" ]; then
+    :
+  else
+    die "Unsupported package method $GIMG_PKG_METHOD"
   fi
 
   _FN_RET[0]="$dir_installation"
@@ -238,7 +270,7 @@ function runner_create()
     :
 	END
 
-  if [ "${GIMG_PKG_TYPE}" = "overlayfs" ]; then
+  if [ "${GIMG_PKG_METHOD}" = "overlayfs" ]; then
     { sed -E 's/^\s+://' | tee -a AppDir/AppRun | sed -e 's/^/-- /'; } <<-END
     :# Unmount after appimage unmounts squashfs
     :function _exit() { pkill -f "overlayfs.*\$WINEPREFIX"; }
@@ -263,7 +295,7 @@ function runner_create()
     :  -o lowerdir="\$WINEPREFIX_RO",upperdir="\$WINEPREFIX_RW",workdir="\$CFGDIR" "\$WINEPREFIX"
     :
 		END
-  elif [ "${GIMG_PKG_TYPE}" = "unionfs" ]; then
+  elif [ "${GIMG_PKG_METHOD}" = "unionfs" ]; then
     { sed -E 's/^\s+://' | tee -a AppDir/AppRun | sed -e 's/^/-- /'; } <<-END
     :# Unmount after appimage unmounts squashfs
     :function _exit() { pkill -f "unionfs.*\$WINEPREFIX"; }
@@ -279,7 +311,7 @@ function runner_create()
     :"\$APPDIR/usr/bin/unionfs" -o uid="\$(id -u)",gid="\$(id -g)" -ocow "\$WINEPREFIX_RW"=RW:"\$WINEPREFIX_RO"=RO "\$WINEPREFIX"
     :
 		END
-  elif [ "${GIMG_PKG_TYPE}" = "readonly" ]; then
+  elif [ "${GIMG_PKG_METHOD}" = "copy" ]; then
     { sed -E 's/^\s+://' | tee -a AppDir/AppRun | sed -e 's/^/-- /'; } <<-END
     :# Copy prefix to outside of appimage
     :if [ ! -d "\$WINEPREFIX" ]; then
@@ -384,6 +416,169 @@ function runner_create()
   read -r
 }
 
+function runner_create_flatimage()
+{
+  # Build Dir
+  local dir_build="$1"; shift
+
+  # Application name, space separated, uppercase
+  local name="$1"; shift
+  declare -a name_upper
+  for i in $name; do name_upper+=("${i^}"); done
+  name="${name_upper[*]}"
+
+  # Binary directory path under c: drive
+  path_install="$1"
+
+  # Binary path under AppDir
+  local path_exec
+  path_exec="$2"
+  path_exec="${path_exec##*AppDir/app/wine/}"
+
+  # Parse yaml
+  "$BIN_PKG" fim-root cp "${GIMG_SCRIPT_DIR}/yq" "\$FIM_DIR_STATIC"
+
+  # Launcher
+  "$BIN_PKG" fim-root cp "${GIMG_SCRIPT_DIR}/launcher-shared" "\$FIM_DIR_STATIC/gameimage-launcher"
+
+  # Create runner script
+  { sed -E 's/^\s+://' | tee AppDir/gameimage.sh | sed -e 's/^/-- /'; } <<-END
+   :set -e
+   :
+   :# Include wine in PATH
+   :PATH="/opt/wine/bin:\$PATH"
+   :
+   :export LC_ALL="\${LC_ALL:-C}"
+   :
+   :# Exports
+   :export DIR_CALL="\$FIM_DIR_BINARY"
+   :export BIN_WINE="/fim/wine.sh"
+   :
+   :# Define configuration directory path
+   :export CFGDIR="\$FIM_DIR_BINARY/.\$FIM_FILE_BINARY.config"
+   :mkdir -p "\$CFGDIR"
+   :
+	END
+
+  if [[ "$GIMG_PKG_METHOD" = "prefix" ]]; then
+    { sed -E 's/^\s+://' | tee -a AppDir/gameimage.sh | sed -e 's/^/-- /'; } <<-END
+     :# Export prefix
+     :export WINEPREFIX; WINEPREFIX="\$CFGDIR/wine"
+     :
+     :# Requires pre-existing prefix to start
+     :[ ! -d "\$WINEPREFIX" ] && { echo "Requires pre-existing prefix to start"; exit 1; }
+     :
+		END
+  elif [ "$GIMG_PKG_METHOD" = "copy" ]; then
+    { sed -E 's/^\s+://' | tee -a AppDir/gameimage.sh | sed -e 's/^/-- /'; } <<-END
+    :# Export prefix
+    :export WINEPREFIX; WINEPREFIX="\$CFGDIR/wine"
+    :
+    :# Copy prefix to outside of appimage
+    :if [ ! -d "\$WINEPREFIX" ]; then
+    :  mkdir -p "\$CFGDIR"/wine
+    :  cp -r "/prefix"/. "\$CFGDIR"/wine
+    :fi
+    :
+    :# Create/Update symlink to the application directory
+    :ln -T -sfn "/rom/$path_install" "\$CFGDIR/wine/drive_c/$path_install"
+    :
+		END
+  else # overlayfs, dynamic
+    { sed -E 's/^\s+://' | tee -a AppDir/gameimage.sh | sed -e 's/^/-- /'; } <<-END
+     :# Export prefix
+     :export WINEPREFIX; WINEPREFIX="\$(readlink -f "/prefix")"
+     :
+		END
+  fi
+
+
+  { sed -E 's/^\s+://' | tee -a AppDir/gameimage.sh | sed -e 's/^/-- /'; } <<-END
+   :# Enter the main executable's directory
+   :cd "\$(dirname "\$WINEPREFIX/$path_exec")"
+   :
+   :# Name of the main executable (without path)
+   :export GIMG_DEFAULT_EXEC="\${GIMG_DEFAULT_EXEC:-$(basename "$path_exec")}"
+   :
+   :# Check YAML integrity
+   :YAML="\$CFGDIR/config.yml"
+   :if ! yq --exit-status 'tag == "!!map" or tag == "!!seq"' "\$YAML" &>/dev/null; then
+   :  echo "cmd: \"{wine} {exec}\"" > "\$YAML"
+   :  echo "runner: ''" >> "\$YAML"
+   :  echo "runner_default: \"true\"" >> "\$YAML"
+   :fi
+   :
+   :# Change startup command
+   :if [[ "\$1" =~ --gameimage-cmd=(.*) ]]; then
+   :  # Define custom command
+   :  CMD="\${BASH_REMATCH[1]}"
+   :  CMD="\${CMD//\"/\\\\\"}" # Escape quotes
+   :  yq -i ".cmd = \"\$CMD\"" "\$YAML"
+   :  exit 0
+   :fi
+   :
+   :# Change runner
+   :if [[ "\$1" =~ --gameimage-runner=(.*) ]]; then
+   :  # Define custom runner
+   :  RUNNER="\${BASH_REMATCH[1]}"
+   :  RUNNER="\${RUNNER//\"/\\\\\"}" # Escape quotes
+   :  yq -i ".runner = \"\$RUNNER\"" "\$YAML"
+   :  exit 0
+   :fi
+   :
+   :# Parse runner location
+   :if { RUNNER="\$(yq '.runner | select(.!=null)' "\$YAML")"; [ -n "\$RUNNER" ]; }; then
+   :  BIN_WINE="\$RUNNER"
+   :fi
+   :
+   :# Parse startup command
+   :if { CMD="\$(yq '.cmd | select(.!=null)' "\$YAML")"; [ -n "\$CMD" ]; }; then
+   :  # Run custom command, replaces {wine} and {exec} strings
+   :  CMD="\${CMD//\{wine\}/\"\$BIN_WINE\"}"
+   :  CMD="\${CMD//\{exec\}/\"\$GIMG_DEFAULT_EXEC\"}"
+   :  CMD="\${CMD//\{here\}/\"\$DIR_CALL\"}"
+   :else
+   :  echo "Startup command is empty, try to erase YAML"
+   :  exit 1
+   :fi
+   :
+   :# Start application
+   :if [ -z "\$GIMG_LAUNCHER_DISABLE" ]; then
+   :  export GIMG_CONFIG_FILE="\$CFGDIR/config.yml"
+   :  export GIMG_LAUNCHER_NAME="$name"
+   :  export GIMG_LAUNCHER_IMG="/fim/desktop-integration/icon.png"
+   :  export GIMG_LAUNCHER_CMD="\$CMD"
+   :  export GIMG_LAUNCHER_EXECUTABLES="\$(find . -iname '*.exe' -exec echo -n '{}|' \;)"
+   :  gameimage-launcher
+   :else
+   :  if [ -z "\$BIN_WINE" ]; then
+   :   echo "-- Wine runner is missing"
+   :   echo "-- You can set it with './\$(basename "$FIM_FILE_BINARY") --gameimage-runner=/path/to/wine'"
+   :   echo "-- The path must be absolute"
+   :   exit 1
+   :  fi
+   :  eval "\$CMD"
+   :fi
+	END
+
+  # Allow execute
+  chmod +x AppDir/gameimage.sh
+
+  # Move into flatimage
+  "$BIN_PKG" fim-exec mv "$DIR_BUILD/AppDir/gameimage.sh" "/fim/gameimage.sh"
+
+  # Erase symlinks from wine user home
+  for i in "$WINEPREFIX/drive_c/users/$(whoami)"/*; do
+    if [ -L "$i" ]; then rm "$i"; fi
+  done
+
+  # Set default command
+  "$BIN_PKG" fim-cmd "/fim/gameimage.sh"
+
+  msg "%b" "Runner script written, make further changes to it if you desire, then press enter..."
+  read -r
+}
+
 function main()
 {
   # Validate params
@@ -394,16 +589,34 @@ function main()
   local cover="${_FN_RET[4]}"
 
   # Create & cd in build dir
-  local dir_build="$(dir_build_create "$dir_src")"
-  cd "$dir_build"
+  export DIR_BUILD="$(dir_build_create "$dir_src")"
+  cd "$DIR_BUILD"
+
+  # Revert prefix
+  if [[ -d "AppDir/app/rom" ]]; then
+    mv -vf "AppDir/app/rom"/* "AppDir/app/wine/drive_c/"
+    rmdir "AppDir/app/rom"
+  fi
+
+  # Revert copy GIMG_PKG_METHOD
+  if [[ -d "$DIR_BUILD/.${name}.${GIMG_PKG_TYPE}.config/wine" ]]; then
+    # If it exists, just erase the one from the config dir
+    if [[ -d "$DIR_BUILD/AppDir/app/wine" ]]; then
+      msg "Remove $DIR_BUILD/.${name}.${GIMG_PKG_TYPE}.config/wine"
+      rm -rf "$DIR_BUILD/.${name}.${GIMG_PKG_TYPE}.config/wine"
+    else
+      msg "Move '$DIR_BUILD/.${name}.${GIMG_PKG_TYPE}.config/wine' to '$DIR_BUILD/AppDir/app'"
+      mv -f "$DIR_BUILD/.${name}.${GIMG_PKG_TYPE}.config/wine" "$DIR_BUILD/AppDir/app"
+    fi
+  fi
 
   export WINEPREFIX="$(pwd)/AppDir/app/wine"
   export BIN_WINE="$(pwd)/AppDir/usr/bin/wine"
   export BIN_WINETRICKS="$(pwd)/AppDir/usr/bin/winetricks"
 
   # Check if wine was moved by the build stage
-  if [ -f "$dir_build"/wine ]; then
-    mv "$dir_build"/wine "$BIN_WINE" 
+  if [ -f "$DIR_BUILD"/wine ]; then
+    mv "$DIR_BUILD"/wine "$BIN_WINE" 
   fi
 
   dir_appdir_create
@@ -428,28 +641,45 @@ function main()
       ;;
 
       test)
-        wine_test "$dir_build/AppDir/app/wine"
+        wine_test "$DIR_BUILD/AppDir/app/wine"
       ;;
 
       build)
         # Select main executable
-        wine_executable_select "$dir_build" "$name"
+        local path_executable="$(wine_executable_select "$DIR_BUILD" "$name")"
 
-        # Create runner script
-        runner_create "$dir_build" "$name" "${_FN_RET[0]}" "${_FN_RET[1]}"
+        # Adjust tree structure based on package method
+        wine_package_method "$path_executable" "$DIR_BUILD" "$name"
+        local dir_installation="${_FN_RET[0]}"
+        local basename_executable="${_FN_RET[1]}"
 
-        # Copy cover
-        ./imagemagick "$cover" "AppDir/${name}.png"
-
-        # Create desktop entry
-        desktop_entry_create "$name"
-
-        # Build appimage
-        appdir_build
-
-        # Rename AppImage file
-        [ -f "${name}.AppImage" ] && rm "${name}.AppImage"
-        mv ./*.AppImage "${name// /-}.AppImage"
+        if [[ "$GIMG_PKG_TYPE" = "flatimage" ]]; then
+          # Copy cover
+          "$BIN_WINE" fim-exec mkdir -p /fim/desktop-integration
+          "$BIN_WINE" fim-exec cp "$cover" /fim/desktop-integration/icon.png
+          # Define path to release package
+          export BIN_PKG="$DIR_BUILD/${name}.${GIMG_PKG_TYPE}"
+          # Copy wine to build dir
+          cp "$BIN_WINE" "$BIN_PKG"
+          # Compress & include prefix 
+          build_flatimage
+          # Create runner script
+          runner_create_flatimage "$DIR_BUILD" "$name" "$dir_installation" "$basename_executable"
+        elif [[ "$GIMG_PKG_TYPE" = "appimage" ]]; then
+          # Create runner script
+          runner_create "$DIR_BUILD" "$name" "$dir_installation" "$basename_executable"
+          # Copy cover
+          ./imagemagick "$cover" "AppDir/${name}.png"
+          # Create desktop entry
+          desktop_entry_create "$name"
+          # Build appimage
+          build_appimage
+          # Rename AppImage file
+          [ -f "${name}.AppImage" ] && rm "${name}.AppImage"
+          mv ./*.AppImage "${name// /-}.AppImage"
+        else
+          die "Invalid package type '$GIMG_PKG_TYPE'"
+        fi
       ;;
     esac
